Contributions from GeoMesa back to SFCurve
* Moved API into api module
* Moved hilbert and zorder curves into their own modules
  - Helps minimize dependency propagation e.g. uzaygezen
* Added SPI access pattern for 2D curve
