+++ api/src/main/scala/org/locationtech/sfcurve/SpaceFillingCurve2D.scala
