+++ project/plugins.sbt
@@ -3,5 +3,5 @@ resolvers ++= Seq(
   Opts.resolver.sonatypeReleases
 )
 
-addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.11.2")
+addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.14.1")
