+++ zorder/src/main/resources/META-INF/services/org.locationtech.sfcurve.SpaceFillingCurveProvider
@@ -1,0 +1,1 @@
+org.locationtech.sfcurve.zorder.ZOrderSFCProvider
