+++ zorder/src/test/scala/org/locationtech/sfcurve/zorder/Z3Spec.scala
